refactor: remove legacy Meson build files, setup cruft, and organize DSP source files into Source/DSP/
