+++ Source/DSP/signal_crossfade.h
