--- setup/win64.ini
@@ -1,22 +1,0 @@
-[binaries]
-name = 'win64'
-c = 'x86_64-w64-mingw32-gcc'
-cpp = 'x86_64-w64-mingw32-g++'
-ar = 'x86_64-w64-mingw32-gcc-ar'
-nm = 'x86_64-w64-mingw32-gcc-nm'
-ld = 'x86_64-w64-mingw32-ld'
-strip = 'x86_64-w64-mingw32-strip'
-pkgconfig = 'pkg-config'
-exe_wrapper = 'wine64'
-
-[host_machine]
-system = 'windows'
-cpu_family = 'x86_64'
-cpu = 'x86_64'
-endian = 'little'
-
-[properties]
-needs_exe_wrapper = true
-
-[libspecbleach:built-in options]
-default_library = 'static'
