--- subprojects/libspecbleach.wrap
@@ -1,3 +1,0 @@
-[wrap-git]
-url = https://github.com/lucianodato/libspecbleach.git
-revision = 091f971c674617966024e7dae00abc6fc6d2475f
